+++ src/GitHubUpdater.py
@@ -9,6 +9,7 @@ import uasyncio as asyncio
 import urequests
 import uos
 import mip
+from ConfigManager import ConfigManager
 
 class GitHubUpdater:
     """Provides online update functionality for device firmware using MIP."""
@@ -18,9 +19,54 @@ class GitHubUpdater:
         self.update_interval = update_interval
         self.auto_reboot = auto_reboot
 
+        self.config_directory = "/config"
+        self.config_file = "secureme.conf"
+        self.network_config_file = "network_config.conf"
+
+        self.pushover_app_token = None
+        self.pushover_api_key = None
+        self.system_status_notifications = None
+        self.general_notifications = None
+        self.security_code_notifications = None
+        self.web_interface_notifications = None
+        self.update_notifications = None
+
         self.headers = {"User-Agent": "GoatGitHubUpdater/1.1"}
         self.latest_version = None
         self.files_to_download = []
+
+    async def initialize(self):
+        """Initializes the server by loading configuration data."""
+        self.config = ConfigManager(self.config_directory, self.config_file)
+        await self.config.read_async()
+
+        self.pushover_app_token = self.config.get_entry("pushover", "app_token")
+        self.pushover_api_key = self.config.get_entry("pushover", "api_key")
+        self.system_status_notifications = self.config.get_entry("pushover", "system_status_notifications")
+        if not isinstance(self.system_status_notifications, bool):
+            self.system_status_notifications = True
+            self.config.set_entry("pushover", "system_status_notifications", self.system_status_notifications)
+            await self.config.write_async()
+        self.general_notifications = self.config.get_entry("pushover", "general_notifications")
+        if not isinstance(self.general_notifications, bool):
+            self.general_notifications = True
+            self.config.set_entry("pushover", "general_notifications", self.general_notifications)
+            await self.config.write_async()
+        self.security_code_notifications = self.config.get_entry("pushover", "security_code_notifications")
+        if not isinstance(self.security_code_notifications, bool):
+            self.security_code_notifications = True
+            self.config.set_entry("pushover", "security_code_notifications", self.security_code_notifications)
+            await self.config.write_async()
+        self.web_interface_notifications = self.config.get_entry("pushover", "web_interface_notifications")
+        if not isinstance(self.web_interface_notifications, bool):
+            self.web_interface_notifications = True
+            self.config.set_entry("pushover", "web_interface_notifications", self.web_interface_notifications)
+            await self.config.write_async()
+        self.update_notifications = self.config.get_entry("pushover", "update_notifications")
+        if not isinstance(self.update_notifications, bool):
+            self.update_notifications = True
+            self.config.set_entry("pushover", "update_notifications", self.update_notifications)
+            await self.config.write_async()
 
     def isNetworkConnected(self):
         """Check if the network interface is connected."""
@@ -29,6 +75,95 @@ class GitHubUpdater:
             return sta.isconnected()
         except Exception:
             return False
+
+    async def send_pushover_notification(self, title="Goat - SecureMe", message="Testing", priority=0, timeout=5):
+        """Send push notifications using Pushover.
+
+            Args:
+            - title: The title for the notification.
+            - message: The message to send.
+            - priority: The notification priority (0-2).
+            - timeout: The request timeout in seconds.
+            """
+        await asyncio.sleep(0)
+
+        url = "https://api.pushover.net/1/messages.json"
+
+        self.pushover_app_token = self.config.get_entry("pushover", "app_token")
+
+        if not self.pushover_app_token:
+            print("A Pushover app token is required to send push notifications.")
+            return
+
+        self.pushover_api_key = self.config.get_entry("pushover", "api_key")
+
+        if not self.pushover_api_key:
+            print("A Pushover API key is required to send push notifications.")
+            return
+
+        data_dict = {
+            "token": self.pushover_app_token,
+            "user": self.pushover_api_key,
+            "message": message,
+            "priority": priority,
+            "title": title
+        }
+        data = self.urlencode(data_dict).encode("utf-8")
+        headers = {"Content-Type": "application/x-www-form-urlencoded"}
+
+        for attempt in range(3):  # Retry up to 3 times
+            try:
+                print(f"Attempt {attempt + 1}: Sending notification...")
+                response = urequests.post(url, data=data, headers=headers, timeout=timeout)
+
+                if response.status_code == 200:
+                    print("Notification sent successfully!")
+                    print("Response:", response.text)
+                    return
+                else:
+                    print(f"Failed to send notification. Status code: {response.status_code}")
+            except Exception as e:
+                print(f"Error sending notification (Attempt {attempt + 1}): {e}")
+            finally:
+                if 'response' in locals():
+                    response.close()
+                await asyncio.sleep(0.5)  # Slight delay before retrying
+
+        print("All attempts to send notification failed.")
+
+    async def send_system_status_notification(self, status_message):
+        """Sends a system status notification via Pushover.
+
+        Args:
+        - status_message: The message to send.
+        """
+        await asyncio.sleep(0)
+
+        if not status_message:
+            print("A status message is required.")
+            return
+
+        try:
+            self.pushover_app_token = self.config.get_entry("pushover", "app_token")
+
+            if not self.pushover_app_token:
+                return
+
+            self.pushover_api_key = self.config.get_entry("pushover", "api_key")
+
+            if not self.pushover_api_key:
+                return
+
+            self.system_status_notifications = self.config.get_entry("pushover", "system_status_notifications")
+
+            if self.system_status_notifications:
+                asyncio.create_task(self.send_pushover_notification(message=status_message))
+        except Exception as e:
+            print(f"Unable to send system status notification: {e}")
+
+    def urlencode(self, data):
+        """Encode a dictionary into a URL-encoded string."""
+        return "&".join(f"{key}={value}" for key, value in data.items())
 
     async def check_for_update(self):
         """Checks for updates from GitHub."""
@@ -123,9 +258,15 @@ class GitHubUpdater:
 
         await self.check_for_update()
         if await self.is_update_available():
+            if self.system_status_notifications:
+                if self.web_interface_notifications:
+                    asyncio.create_task(self.send_system_status_notification(status_message=f"Firmware update available. Updating from {self.current_version} to {self.latest_version}"))
             await self.download_update()
             self.current_version = self.latest_version
             print(f"Firmware update complete. Updated to version {self.current_version}")
+            if self.system_status_notifications:
+                if self.web_interface_notifications:
+                    asyncio.create_task(self.send_system_status_notification(status_message=f"Firmware update complete. Updated to {self.latest_version}"))
             if self.auto_reboot:
                 print("Restarting system...")
                 machine.reset()
@@ -135,6 +276,8 @@ class GitHubUpdater:
     async def run_periodically(self):
         """Periodically update device firmware from GitHub."""
         print("Initializing automatic update...")
+
+        await self.initialize()
 
         await asyncio.sleep(30)  # Delay before starting